BACKEND/requirements.txt
django
djangorestframework
mysql-connector-python
bcrypt